Add validation utils for encrypted file metadata (#25)
When scanning an encrypted file, clients are expected to send the
content scanner some metadata it can use to decrypt it. This is
typically the `file` property of the `m.room.message` event content. See
https://github.com/matrix-org/matrix-content-scanner-python/blob/main/docs/api.md#post-_matrixmedia_proxyunstabledownload_encrypted
to see how it fits into the API, and an example.

This adds a `validate_encrypted_file_metadata` util that performs a few
validation checks on the client-provided data to ensure it's in the
right format before passing it on to the scanner.

Closes #7