Use Futures to deduplicate incoming requests (#37)
Fixes #23 

As of #19, we cache (most) scan results in memory to avoid having to
download and scan the same file twice (with the caveat that we don't
cache the contents of big files, but it still saves scanning them).
However, it does not prevent duplicate work if multiple requests for the
same file happen simultaneously, since we only update this cache when a
scan completes. This is a legitimate concern, since we can easily
imagine a situation where a user posts a large file into a room and
multiple users try to download it at the same time.

As of this change,

* when a request for a given file with given encryption and thumbnailing
parameters comes in, and we're not already running a download/scan for
it (with those parameters), we create a `Future` that we store into a
new `_current_scans` cache, and run the scan. Once the scan completes,
we resolve the `Future` with the result, and remove it from the cache.
* when a request for the same file comes in _before_ the previous
request completed, we don't download and/or scan the file - instead we
look up the matching `Future` from the `_current_scans` cache and
`await` it. We then return its result.